chore: disable blank issues
As a test we are trying non blank issues and enhancements as default.
This rationale is that if the input format is pre definied it will
in the future be easier get oversight and statistics for input.
It will guide the reporter to input better data.

If a project has specific needs, they can add their own settings.
This was also possible before this, so this just sets the new default.

Signed-off-by: Josef Andersson <josef.andersson@digg.se>